Refactored ThreadSafeCache to be a mixin. Changed Cache19 to use Hash#shift in Ruby 2.1+, a performance improvement. Fixed typo in Readme.
