+++ lib/lru_redux.rb
@@ -6,11 +6,13 @@ module LruRedux
 end
 
 require "lru_redux/version"
+require "lru_redux/util"
+
 if LruRedux.is_19?
   require "lru_redux/cache19"
+  require "lru_redux/cache19_legacy" if RUBY_VERSION < "2.1.0"
 else
   require "lru_redux/cache"
 end
+
 require "lru_redux/thread_safe_cache"
-require "lru_redux/thread_safe_cache_jruby" if
-    RUBY_PLATFORM == 'java' && JRUBY_VERSION < '9.0'
