+++ lib/lru_redux/util.rb
@@ -1,0 +1,4 @@
+require 'util/safe_sync'
+
+require 'util/safe_sync_jruby' if
+    RUBY_PLATFORM == 'java' && JRUBY_VERSION < '9.0'
