+++ lib/lru_redux/util/safe_sync_jruby.rb
@@ -1,4 +1,4 @@
-class LruRedux::ThreadSafeCache
+class LruRedux::Util::SafeSync
   def getset(key, &block)
     synchronize do
       super(key, &block)
